third_party/WebKit/LayoutTests/inspector/sources/pretty-print-javascript-9.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/sources-test.js"></script>

<script>

function test()
{
    var testJSFormatter = InspectorTest.testPrettyPrint.bind(InspectorTest, "text/javascript");

    InspectorTest.runTestSuite([
        function parenthesizedExpressions(next)
        {
            var mappingQueries = ["if", "((a))", "((b));", "else", "(c)"];
            testJSFormatter("if((a))((b));else (c);", mappingQueries, next);
        },
    ]);
}

</script>

</head>

<body onload="runTest()">
<p>Verifies JavaScript pretty-printing functionality.</p>
</body>
</html>
